src/requirements.txt
asteval==1.0.6
fastapi==0.115.12
google-genai==1.15.0
litellm==1.70.4
prompt-toolkit==3.0.51
python-dotenv==1.1.0
uvicorn==0.34.2
googlesearch-python